stdaddr: Move from internal/staging to txscript.
This moves the new stdaddr package from the internal staging area to the
txscript module and updates the relevant paths and package README.md
accordingly.
