+++ txscript/stdaddr/address_test.go
