+++ txscript/stdaddr/bench_test.go
