+++ txscript/stdaddr/error.go
