+++ txscript/stdaddr/error_test.go
